Feat: More command editor updates
New command types documented.

Ability to split command arg1 into 8 flags and a byte.

Remove the old list view of commands.

Change the compiling of the map to reverse through the sectors
in an attempt to fix the ordering of the texture mapping array
for the study fireplace animation. Not sure if there's something
wrong with the map compiler or if the vanilla game is dependent on a
chance ordering of the textures.