Make sure face is not set to its own sister when pasting
When pasting, face sisters are recalculated by looking for equal,
opposite vertices. When rotating sectors with very close vertices, this
could result in the vertices overlapping making face.v1 equal to face.v2
which could cause a face to be assigned as its own sister. This adds a
check to prevent that.